BUG: Crash in subtractjs8()
Deals with a frequent crash in subjtractjs8(). Believe this to be due to a shared common block being used for parameterized arrays. In such instances, it’s often down to luck if that’ll actually work; basically, first one in wins, so you’d best hope that’s the largest one. Changed to use allocations.