Simpler memory management in NotificationAudio
QByteArray is an implicitly-shared COW class, and we never change it after we read it into cache, so memory managment can be simpler here. Removed unused decoder instance variable, and moved use of file instance variable to automatic local scope. Converted from QMap to QCache, as order of cached entries is irrelevant to us.