Use std::rotate in detector
Since we have C++17 available to us, std::rotate is available, and on every implementation tested, auto-vectorizes into, for example, use of the XMM registers on Intel machines (note, not in debug mode, the optimizer is required). The result runs at least as fast, typically faster, than the custom versions here, and eliminates use of the static temporary array requirement.