Collapse TX and RX frequencies
Separate TX and RX frequencies is a vestige of WSJTX; in JS8Call, there’s really only a single frequency, which is our offset in hertz into the dial frequency. To date the two separate frequencies have had to have been kept in sync at all times; this change removes the distinction and instead uses a single frequency.