Refactoring; eliminate bandComboBox
This widget is always hidden; it can’t be edited or activated, and thus ends up being, effectively, an alias for m_lastBand. As m_bandEdited can’t ever be anything but false, conditional code relying on it being true will never be executed. The LiveFrequencyValidator was only used by this combo box.